requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
SQLAlchemy==1.4.23
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
